time: use wakelocks in alarmtimer
The wakelock core does not respect wakeup events and cannot be trivially
modified to support them.  If alarmtimer tries to delay suspend for a
pending event, this results in multiple failed attempts to suspend.  If
wakelocks are enabled, use them instead of wakeup events.  While we're
at it, move alarmtimer to suspend_prepare so it can abort suspend
sooner.