msm: kgsl: Fix page align allocation overflow
KGSL does alignments for all user memory requests.  If the
requests is large enough the alignment operation will overflow
resulting in a valid request of size 0.  It's just a matter of
time until the kernel or user tries to then do something useful
with this buffer resulting in a hang.  Perform a sanity check
after doing the alignments to ensure the overflow did not occur.